feat(link): allow to use the daisy pipeline app with the dp2 tool
If no webservice information is defined in the configuration and a valid app_path is set, the tool now tries to use the DAISY Pipeline app.
The tool check for a running instance of the DAISY Pipeline app and forward commands to it if it finds one.
If starting is set to true in configuration, the tool can also launch the daisy pipeline app or the launched pointed by the app_path setting.