Allow using the Pipeline app with the dp2 tool
Check for a running instance of the DAISY Pipeline app, and forward
commands to it. If the `starting` setting is set to true,
automatically launch the app if it is not running yet. The app will in
its turn forward commands back to dp2, passing the right webservice
connection info.