Further improve the default configuration file and help text
Also don't mention the possibility of specifying a webservice launch
script. We are keeping this feature for backward compatibility, but it
is not certain that it will actually be useful, so we keep it a hidden
feature in order to not complicate things for no reason.

Also added a clarifying comment in the code.