Retrieve app path for running pid
Fixes an error due to misunderstanding on my side: I though the go-ps
`process.Executable()` was returning a full path to the executable,
but it only returns the name of the executable.

This commit uses the gopsutil package to retrieve the full path of a
running app's executable using it's PID, as to allow the dp2 tool to
reuse it.

See https://github.com/daisy/pipeline-cli-go/pull/54